requirements.txt
ccxt
pandas
pandas-ta
gspread
oauth2client
python-telegram-bot
schedule
